- Re-structure the header classes - Add documentation - Fix lint and fmt
